requirements/test.txt
ariadne==0.6.0
coverage>=4.4
pytest>=3.3.1
pytest-cov>=2.4.0
pytest-django>=3.1.2
